requirements.txt
absl-py==0.7.1
appdirs==1.4.3
asgiref==3.2.3
astor==0.7.1
attrs==19.1.0
configparser==4.0.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-widget-tweaks
django-chartjs
dnspython==1.16.0
docutils==0.15.2
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jsonschema==3.0.1
matplotlib==3.0.3
mongo==0.2.0
npm==0.1.1
numpy==1.16.2
pandas==0.24.2
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2018.9
PyYAML==5.2
pyzmq==18.0.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==20.0.15
w3lib==1.21.0
widgetsnbextension==3.4.2
